N001/requirements.txt
anyio==3.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.1
fonttools==4.39.3
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
RPi.GPIO==0.7.1
six==1.16.0
smbus==1.1.post2
smbus2==0.4.2
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
zipp==3.15.0
